test(api): split api container tests (#9541)
