+++ packages/amplify-e2e-tests/src/__tests__/containers-api.test.ts
@@ -2,7 +2,6 @@ import {
   addRestContainerApi,
   amplifyConfigureProject,
   amplifyPushWithoutCodegen,
-  amplifyPushSecretsWithoutCodegen,
   createNewProjectDir,
   deleteProject,
   deleteProjectDir,
@@ -12,8 +11,6 @@ import {
 } from 'amplify-e2e-core';
 import fetch from 'node-fetch';
 import { getAWSExports } from '../aws-exports/awsExports';
-import * as fs from 'fs-extra';
-import path from 'path';
 
 async function setupAmplifyProject(cwd: string) {
   await amplifyConfigureProject({
@@ -64,42 +61,4 @@ describe('amplify api add', () => {
     await amplifyPushWithoutCodegen(projRoot);
   });
 
-  it('init project, api container secrets should work', async () => {
-    const envName = 'devtest';
-    const apiName = 'containersecrets';
-    await initJSProjectWithProfile(projRoot, { name: 'multicontainer', envName });
-    await setupAmplifyProject(projRoot);
-    await addRestContainerApi(projRoot, { apiName });
-    await setupContainerSecrets(projRoot, apiName)
-    await amplifyPushSecretsWithoutCodegen(projRoot);
-    const awsExports: any = getAWSExports(projRoot).default;
-    const {
-      aws_cloud_logic_custom: [{ name, endpoint }],
-    } = awsExports;
-    expect(name).toBeDefined();
-    expect(endpoint).toBeDefined();
-
-    const result = await (await fetch(`${endpoint}/password`)).text();
-    expect(result).toEqual('CONTAINER_SECRETS_PASSWORD');
-  });
-
 });
-
-const setupContainerSecrets = async (projRoot: string, apiName: string) => {
-  // Get the api folder path
-  const apiFolder = path.join(projRoot, 'amplify', 'backend', 'api', apiName);
-
-  // Get the resources folder paths
-  const secretsFolder = path.join(__dirname, '..', '..', 'resources', 'api-container', 'secrets');
-  const dockerFile = path.join(__dirname, '..', '..', 'resources', 'api-container', 'docker-compose.yml');
-  const expressFile = path.join(__dirname, '..', '..', 'resources', 'api-container', 'express', 'secret-password-index.js');
-  
-  // Read the docker and express index files from resources
-  const dockerFileContents = await fs.readFile(dockerFile);
-  const expressFileContents = await fs.readFile(expressFile);
-
-  // Write the files on the project folder
-  await fs.copy(secretsFolder, path.join(apiFolder, 'secrets'));
-  await fs.writeFile(path.join(apiFolder, 'src', 'docker-compose.yml'), dockerFileContents);
-  await fs.writeFile(path.join(apiFolder, 'src', 'express', 'index.js'), expressFileContents);
-};
